move _zip_folder_to_byte_stream to fractal_server/zip_tools.py
