+++ fractal_server/app/routes/aux/_job.py
@@ -1,9 +1,5 @@
-import os
-from io import BytesIO
 from pathlib import Path
 from typing import Union
-from zipfile import ZIP_DEFLATED
-from zipfile import ZipFile
 
 from ...models.v1 import ApplyWorkflow
 from ...models.v2 import JobV2
@@ -24,22 +20,3 @@ def _write_shutdown_file(*, job: Union[ApplyWorkflow, JobV2]):
     shutdown_file = Path(job.working_dir) / SHUTDOWN_FILENAME
     with shutdown_file.open("w") as f:
         f.write(f"Trigger executor shutdown for {job.id=}.")
-
-
-def _zip_folder_to_byte_stream(*, folder: str) -> BytesIO:
-    """
-    Get byte stream with the zipped log folder of a job.
-
-    Args:
-        folder: the folder to zip
-    """
-
-    byte_stream = BytesIO()
-    with ZipFile(byte_stream, mode="w", compression=ZIP_DEFLATED) as zipfile:
-        for root, dirs, files in os.walk(folder):
-            for file in files:
-                file_path = os.path.join(root, file)
-                archive_path = os.path.relpath(file_path, folder)
-                zipfile.write(file_path, archive_path)
-
-    return byte_stream
